Multiple improvements for steamdeck (#99)
Paste the SteamgridDB API key in from clipboard automatically if it looks like a API key. This is nice for steam deck so you don't have to do a right-click paste or worse manually write in the key.
BoilR guesses what game it should find images for based on the game name, but it might be wrong, it is now possible to correct this
No longer show a back button on the user select screen (it did nothing)
Now you can change the images for ALL your shortcuts, not just the BoilR ones.